+++ src/web/Learning.Domain/Notification/ModelExamConfiguration.cs
@@ -14,12 +14,12 @@ public class ModelExamConfiguration : DomainBase
     /// <summary>
     /// Score per question
     /// </summary>
-    public int Score { get; set; }
+    public float Score { get; set; }
 
     /// <summary>
     /// Negative score for wrong answer
     /// </summary>
-    public int NegativeScore { get; set; }
+    public float NegativeScore { get; set; }
 
     /// <summary>
     /// Value is seconds
